requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
SQLAlchemy==2.0.20
alembic==1.11.1
pydantic==1.10.7
PyJWT==2.7.0
pytest==7.4.0
python-jose==3.3.0  # Adicionando o python-jose
